+++ maxminddb/reader.py
@@ -81,6 +81,7 @@ class Reader(object):
 
         self._decoder = Decoder(self._buffer, self._metadata.search_tree_size +
                                 self._DATA_SECTION_SEPARATOR_SIZE)
+        self.closed = False
 
     def metadata(self):
         """Return the metadata associated with the MaxMind DB file"""
@@ -181,6 +182,15 @@ class Reader(object):
         # pylint: disable=unidiomatic-typecheck
         if type(self._buffer) not in (str, bytes):
             self._buffer.close()
+        self.closed = True
+
+    def __exit__(self, *args):
+        self.close()
+
+    def __enter__(self):
+        if self.closed:
+            raise ValueError('Attempt to reopen a closed MaxMind DB')
+        return self
 
 
 class Metadata(object):